split level scene into managers
